Weighted average

+++ 0x04-python-more_data_structures/100-weight_average.py
@@ -6,4 +6,4 @@ def weight_average(my_list=[]):
         ml += row[0] * row[1]
         sm += row[1]
 
-    return float(ml / sm)
+    return ml / sm
